Update dependencies
This required removal of QRegExp and full conversion to QRegularExpression.
In 5.x, we needed QRegExp, as the "bootstrap" minimal configuration for
building host tools for cross-compilation did not include the
QRegularExpression components.

One of the tests was also converted from using the Qt::DateFormat enum to
using the Qt::DayOfWeek enum, as several of the used DateFormat enums were
removed in Qt6.

Fixes: QTBUG-72592
Task-number: QTBUG-84590
Change-Id: I3414e6c106407819790de9cd179914ed45ece07e
Reviewed-by: Michael Brasser <michael.brasser@live.com>