Update commercial license headers
Updated header.COMM to the files in tqtc-tqremoteobjects.
Tests, examples or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the line count causes failures in tests.

Task-number: QTQAINFRA-4142
Change-Id: I85e605e9b8cf254a5cb9bee4211b06edb2fe8639
Reviewed-by: Akseli Salovaara <akseli.salovaara@qt.io>