Add a helper class for waiting for model changes in tests
The tests are using complicated logic when waiting for model changes.
This patch introduces a new WaitHelper class, which uses a QPromise to
indicate the start and finish of a task, and QFutureWatcher to wait for
the associated future to finish. This class will be reused in other
parts of code, where waiting is required.

Task-number: QTBUG-90688
Change-Id: I4a412fa555be342c168e526601f8fc56d5091872
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
(cherry picked from commit a290ec98c2edf753f253d22d245190bb19f9822e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>