Don't try to access replica's connection when it's not initialized yet
After the recent changes for supporting different serialization
backends, we need to get the proper codec from the connection for
serializing data. However, it's possible that we will be trying to
serialize data, when the connection is not established, e.g. client
started without a host, and is trying to send heartbeat packets to check
for connection (like in the attached bugreport), or the host is
destroyed while the client is still trying to send data).

Check if the connection with the source is not null, before using it.
Note, that these checks were done also before, but after serializing the
data. Now we do them before.

Fixes: QTBUG-94513
Change-Id: Ie2ca558fc11044513b7886aed736383b16b7464b
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
(cherry picked from commit 6d0d2f922bacbcdb5b482d8272c04ec408dfceaf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>