Provide methods to set the Host codec
Each codec will have different ways of sending packet size, so we need an
agnostic way of passing along the choice of codec.  This is done in the
Handshake, which now passes 3 bytes and a utf-8 string.

Byte 1: Size of the packet (< 255 bytes), excluding the size byte itself.
Byte 2: PacketID enum, must be Handshake, truncated to 1 byte.
Byte 3: A revision number for the Codec (unused in this patch)
String: utf08 encoded name for the Codec.

This also provides a codec manager for each node.  The codec holds a packet
buffer, so each node needs its own instance (but only one instance) of
each type of codec used.  The eventloop ensures we don't have simultaneous
reads/writes to different objects.

Additionally:
A clearCodec() method was added to ExternalIoDevice, so on disconnect and
reconnect, the pipe will be ready to receive a new Handshake packet.
The CodecBase getPayload() method now returns non-const reference, so the
value can be written to, not just read.

Change-Id: Id46e78339ef6814a7b55c524ba1963862a8f75eb
Reviewed-by: Michael Brasser <michael.brasser@live.com>