Refactor decodeVariant to pass by rvalue ref/std::move
The decodeVariant calls were weird in how their parameters were pass by
reference so they could be changed, but without forcing a copy.  It is
clearer to pass by rvalue ref, using std::move on the calling side.
This rippled up to calls that called the method, including setProperties
which is part of the repc generated code.

Because of the repc change, this should *not* be picked to Qt5.

Change-Id: Ieb1b0620569ad8eb9797edc57cc189d0b426510c
Reviewed-by: Sona Kurazyan <sona.kurazyan@qt.io>
(cherry picked from commit f43e2918d05df688c19d3164b43418747d4ac2ef)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>