Include Model/Class info in class signature
This patch addresses the underlying issue where a subclass mismatch was
not detected.

Note: this change is not backwards compatible, so any rep MODEL or CLASS
members will have a different signature in Qt5, so mixing Qt5 and Qt6 will
not run.  Picking to 5.15 would only affect the commercial costumers,
which would be more confusing.

Pick-to: 6.2
Change-Id: Idd40e36285760b8c15398c38defdc58350f85c50
Reviewed-by: Michael Brasser <michael.brasser@live.com>