Restructure CMake API docs to add them to the main CMake docs page
The main CMake command reference page for all modules uses annotated
lists of items that are grouped together (for example, see the link in
the bugreport). It has an entry for QtRemoteObjects, but there's no
group with the name it links to, so the docs don't show up. Moved the
docs for each QtRemoteObjects CMake functions to separate pages, which
are grouped as the main CMake doc page expects. This makes CMake API
docs for QtRemoteObjects consitent with the docs for other modules.

Task-number: QTBUG-95433
Change-Id: Ia6ebe1a825ec9bcdfb4628b92a32ad31d3938e24
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit b62f689c2dffdcd648befad4449bdded76da246c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>