Support general QMap<*>/QHash<*> types with a proxy
Extends the QList<*> fix to work with associative containers as well.

Note, this doesn't support enums as keys (to be added in a future commit)

There are two use-cases which should be supported by a fix.
1) Proxy, where the parameter is stored, but then forward to downstream
   replicas which can be compiled.  In this case, the ransformation needs
   to support conversion back to the declared type.
2) Dynamic replicas being used from QML.  In this case, the dynamic type
   should be made available to QML in a way that is equivalent to the
   declared type.

Dynamic storage/QML support are from the type QtROAssociativeContainer.
This class is derived from QVariantMap, but stores a few extra members to
allow forwarding without losing type information.

Over-the-wire transfer is via a new class (QAS_), which stores elements
in a QByteArray based on the valueType, not the container.  Since the type
will always be fully known at the source, we can use metatype detection to
separate out the problematic types (an associative container of Q_GADGET
types), and convert those to QAS_ containers for sending.

Note: there is one additional piece needed.  When we send the definition
of a dynamic type (needed for the creation of the dynamic metaobject), we
need to change the type of any member or parameter that is a such a
container to QtROAssociativeContainer.

Change-Id: Ib50a01b4f9a3ecaa25a5a933feb6ed843e104b15
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
(cherry picked from commit 4bf86077a1377bf6babfd362a3e8e546153d8973)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>