Move class signature handling to parser.g
This refactors the signature handling so the AST knows the signatures of
the included types.  This, in turn, allows the parser headers to be used
directly for interpreting .rep files, without needing to includes pieces
from the repc tool itself just for the signature.


This also augments the signatures to use more fields
* Handle individual elements of template parameters.  For instance,
  "QHash<flag, POD>" will have a signature based on the flag's signature,
  the POD's signature and the string "QHash".
* Flags include the signature of the enum they are based on
* Enum signatures will include the type (if specified) and scope.

Pick-to: 6.2
Change-Id: Id4e68e61a1c2690939e368e1e8e7555eb81fd68a
Reviewed-by: Michael Brasser <michael.brasser@live.com>