Get rid of calls to virtual methods inside constructors
Move the implementations of setRegistryUrl() and setHostUrl() virtual
methods to the non-virtual methods in the corresponding private classes
and call them from the constructors. Although in the context of the
current code the virtual method calls were legal and didn't lead to
anything unexpected, but generally it's not recommended (so the
codechecker complains) and can lead to issues in future, for example
when refactoring or changing the code.

Pick-to: 6.2
Task-number: QTBUG-95752
Change-Id: I13b97e3a1e179b0e5ce5bc7e9bbb9f5bb7166ebb
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>
Reviewed-by: Michael Brasser <michael.brasser@live.com>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>