Allow PODs to include enums/flags
The goal is to allow a new syntax:
POD myPOD{
    ENUM MyEnum {FOO, BAR}
    MyEnum myEnum,
    QString myString
}, in addition to the current syntax
POD myPOD(QString myString)

The challenge is that the parsing simplified the detection of parameter
types and parameter names by grabbing everything between the parentheses
and having a separate chunk of code split that up.  The new POD syntax
requires the parser to detect and handle parameter types (including
templated types) and handle that syntax.

This converts the enum parsing to use new symbol and value tokens,
rather than the enum_param token, which is necessary to enable a more
generic type detection.

The pod2 type has comparable syntax to the previous version, but uses
brackets ('{' and '}') instead of parentheses for the definition.

While this code detects the parameter type and parameter name distinctly,
it currently merges all of the parameter content into a single string for
processing using the original postprocess method.  This ensures the capture
is correct and there aren't any regressions.

The tests were modified for the new POD syntax.

The new code normalizes whitespace, rather than keeping the source
formatting.

Change-Id: I671b23852fc2b515bcaea1ed389cfbde122683bd
Reviewed-by: Michael Brasser <michael.brasser@live.com>
(cherry picked from commit 045cbcbc50d51640142691c8cd78dd868e03ba1b)