Fix metatype registration of signal-slot params of type QList/QMap/QHash
When declaring signals/slots that take QList/QMap/QHash parameters, we
register metatypes for them. However, we don't always register metatypes
for the parameter types of those containers. In particular, if a
parameter is declared as a POD type in a separate .rep file, metatype
registration for it is skipped.

Fixed the code generation for metatype registration for signal-slot
parameters of type QList/QMap/QHash to also register metatypes for the
contained types if required.

Fixes: QTBUG-97704
Change-Id: Ic81895815c380e841c04f2d1400fc2c2cf0e3814
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
(cherry picked from commit 43dfcd1d1a03b4243b45ddfc367b93f07df8e84b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>