Fix test failures on macOS-arm64
There seems to be some slowness for some of macOS-arm64 configurations
on CI, which causes failures.

- Fix tst_usertypes by increasing the test timeouts.

- Fix the client of tst_Integration_External to check the state of the
replica immediately after the stateSpy count changes, to not "loose" the
intermediate state change to "Suspect". Additionally, add a delay
between the state changes on the server-side, to make sure the client
will manage to observe the state changes, and debug prints for easier
debugging in case of failures.

Pick-to: 6.3 6.2
Fixes: QTBUG-99269
Change-Id: I22e5eeae51421fe9b7419ffcec887c41e1912f5c
Reviewed-by: Jani Heikkinen <jani.heikkinen@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>