Update commercial license headers
Updated header.COMM to the files in tqtc-qtremoteobjects.
Examples, tests, or documentation files are not updated.

The commercial license header may contain some
additional lines so that its line count equals with
the earlier license header. Reason for this is that
some autotests use hard coded line numbers and a
change in the the line count causes failures in tests.

Task-number: QTQAINFRA-4933
Change-Id: Ie2f4121c93da8998cb47b91e88e79e4bfd3cb043
Reviewed-by: Sona Kurazyan <sona.kurazyan@qt.io>