Fix missing QIODevice disconnect() signal connection
The QtROExternalIoDevice fails to register a disconnect()
signal from the underlying QIODevice.  If the server tries
to initiate communication after the client disconnects,
it will segfault trying to access the QIODevice.

There is a metaObject check for the disconnect() signal before
connecting, as disconnect() isn't part of the QIODevice
interface.  However, metaObject::indexOfSignal is being passed
a SIGNAL(disconnect()), which is not the correct format. It
should be passed the normalized signature "disconnect()"

Change-Id: I6cc6d51faff49d4106f4126352e1e9e6cf6165a2
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>
(cherry picked from commit d4414cb7baa8a0049a51d5ef4a485cf6ae5d65db)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>