Fix missing QIODevice disconnect() signal connection
The QtROExternalIoDevice fails to register a disconnect()
signal from the underlying QIODevice.  If the server tries
to initiate communication after the client disconnects,
it will segfault trying to access the QIODevice.

There is a metaObject check for the disconnect() signal before
connecting, as disconnect() isn't part of the QIODevice
interface.  However, metaObject::indexOfSignal is being passed
a SIGNAL(disconnect()), which is not the correct format. It
should be passed the normalized signature "disconnect()"

Pick-to: 5.15 6.2 6.4
Change-Id: I6cc6d51faff49d4106f4126352e1e9e6cf6165a2
Reviewed-by: Brett Stottlemyer <bstottle@ford.com>