Move qmlmodelviewclient example to tests/manual/examples
This QML example does not work properly. It only shows a single
column because it uses ListView. If it is reworked to use TreeView,
there are still a lot of issues. Like that setBackground on the
server side sets the background role to a QBrush, but only QColor is
currently supported in QML.

Task-number: QTBUG-112850
Pick-to: 6.5
Change-Id: Ib3224052d1a25bcece31f3ea69c01713f422ef22
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>