QNX QIODevice subclasses: remove Q_DECLARE_TYPEINFOs
QIODevice subclasses are neither copy- nor movable, so they shouldn't
be marked as trivially relocatable, either. I have a QtCore change
lined up that statically asserts that relocatable types are actually
copyable or movable, and this is holding it up.

It's a public class, though, so we need to be mindful of SC and
BC:

For SC, Q_DECLARE_TYPEINFO never really applied. Yes, you can
construct situations where a user statically asserted that this type
was trivially relocatable, but that's SiC Type A: they just need to
remove the assertion.

For BC, Qt no longer has types that change layout based on QTypeInfo
(hello, Q5List), but user code might, but that's their problem. If
existing code used trivial relocatability (memmove instead of move +
destroy), it can continue to do so, no change there. New code will not
have that optimization, but they're equivalent, as long as relocations
happen atomically in one or the other way. Qt doesn't leak which type
of relocation is chosen between functions (by passing a bool parameter
to a function, e.g.); any use of QTypeInfo is contained within the
function that uses it. If users leak it, that's again their problem.
But we should document it, so:

[ChangeLog][QtRemoteObjects][Potentially Source-Incompatible
Changes][QQnxNativeIo/QIOQnxSource] These classes are no longer marked
as Q_RELOCATABLE_TYPE.

Fixes: QTBUG-119870
Change-Id: I08c082a403433dc571b8c6ab337bdc717919c8b9
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>