CMake: Make qtremoteobjects tests standalone projects
Add the boilerplate standalone test prelude to each test, so that they
can be opened with an IDE without the qt-cmake-standalone-test script,
but directly with qt-cmake or cmake.

Boilerplate was added using the following scripts:
https://git.qt.io/alcroito/cmake_refactor

Manual adjustments were made where the code was inserted in the wrong
location.

Pick-to: 6.8 6.9 6.10
Task-number: QTBUG-93020
Change-Id: Ib2fd9fe430c8215dbbfd8c91964ac1b60e0d570b
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>